backend: change the landType logic
Now there is not the landType attribute in the land object. It is  now

+++ backend/src/main/java/com/toprakrehberi/backend/controllers/LandTypeController.java
@@ -28,7 +28,7 @@ public class LandTypeController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<LandTypeDTO> getLandTypeById(@PathVariable("id") Long id) {
+    public ResponseEntity<LandTypeDTO> getLandTypeById(@PathVariable("id") int id) {
         LandType landType = landTypeService.getLandTypeById(id);
         if (landType != null) {
             return new ResponseEntity<>(convertToDTO(landType), HttpStatus.OK);
