+++ backend/src/main/java/com/toprakrehberi/backend/repositories/LandTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LandTypeRepository extends JpaRepository<LandType, Long> {
+public interface LandTypeRepository extends JpaRepository<LandType, Integer> {
 }
